+++ pkgs/native_assets_builder/test/build_runner/helpers.dart
@@ -46,6 +46,7 @@ Future<BuildResult> build(
   int? targetMacOSVersion,
   int? targetAndroidNdkApi,
   Target? target,
+  bool linkingEnabled = false,
 }) async =>
     await runWithLog(capturedLogs, () async {
       final result = await NativeAssetsBuildRunner(
@@ -64,6 +65,7 @@ Future<BuildResult> build(
         targetIOSVersion: targetIOSVersion,
         targetMacOSVersion: targetMacOSVersion,
         targetAndroidNdkApi: targetAndroidNdkApi,
+        linkingEnabled: linkingEnabled,
       );
 
       if (result.success) {
@@ -155,6 +157,7 @@ Future<(BuildResult, LinkResult)> buildAndLink(
         targetIOSVersion: targetIOSVersion,
         targetMacOSVersion: targetMacOSVersion,
         targetAndroidNdkApi: targetAndroidNdkApi,
+        linkingEnabled: true,
       );
 
       if (!buildResult.success) {
@@ -217,6 +220,7 @@ Future<BuildDryRunResult> buildDryRun(
   bool includeParentEnvironment = true,
   List<String>? capturedLogs,
   PackageLayout? packageLayout,
+  required bool linkingEnabled,
 }) async =>
     runWithLog(capturedLogs, () async {
       final result = await NativeAssetsBuildRunner(
@@ -228,6 +232,7 @@ Future<BuildDryRunResult> buildDryRun(
         workingDirectory: packageUri,
         includeParentEnvironment: includeParentEnvironment,
         packageLayout: packageLayout,
+        linkingEnabled: linkingEnabled,
       );
       return result;
     });