+++ pkgs/native_assets_builder/test/build_runner/link_dry_run_test.dart
@@ -30,6 +30,7 @@ void main() async {
           packageUri,
           logger,
           dartExecutable,
+          linkingEnabled: true,
         );
         expect(buildResult.assets.length, 0);
 
@@ -70,6 +71,7 @@ void main() async {
           packageUri,
           logger,
           dartExecutable,
+          linkingEnabled: true,
         );
         expect(buildResult.success, true);
         expect(_getNames(buildResult.assets), orderedEquals(builtHelperAssets));