+++ pkgs/native_assets_builder/test/build_runner/resources_test.dart
@@ -33,6 +33,7 @@ void main() async {
           packageUri,
           logger,
           dartExecutable,
+          linkingEnabled: true,
         );
 
         Iterable<String> buildFiles() => Directory.fromUri(