+++ pkgs/native_assets_builder/test_data/complex_link/hook/build.dart
@@ -17,6 +17,6 @@ void main(List<String> args) async => build(args, (config, output) async {
             package: packageName,
           ),
         ),
-        linkInPackage: 'complex_link',
+        linkInPackage: config.linkingEnabled ? 'complex_link' : null,
       );
     });