+++ pkgs/native_assets_cli/test/api/build_test.dart
@@ -54,6 +54,7 @@ void main() async {
       ),
       buildMode: BuildMode.release,
       linkModePreference: LinkModePreference.preferDynamic,
+      linkingEnabled: false,
     );
     final configJson = (config1 as BuildConfigImpl).toJsonString();
     buildConfigUri = tempUri.resolve('build_config.json');