+++ pkgs/native_toolchain_c/test/cbuilder/cbuilder_build_failure_test.dart
@@ -42,6 +42,7 @@ void main() {
         envScript: envScript,
         envScriptArgs: envScriptArgs,
       ),
+      linkingEnabled: false,
     );
     final buildOutput = BuildOutput();