+++ pkgs/native_toolchain_c/test/cbuilder/cbuilder_cross_macos_host_test.dart
@@ -58,6 +58,7 @@ void main() {
             linkModePreference: linkMode == DynamicLoadingBundled()
                 ? LinkModePreference.dynamic
                 : LinkModePreference.static,
+            linkingEnabled: false,
           );
           final buildOutput = BuildOutput();
 
@@ -138,6 +139,7 @@ Future<Uri> buildLib(
     linkModePreference: linkMode == DynamicLoadingBundled()
         ? LinkModePreference.dynamic
         : LinkModePreference.static,
+    linkingEnabled: false,
   );
   final buildOutput = BuildOutput();