+++ pkgs/native_toolchain_c/test/cbuilder/cbuilder_cross_windows_host_test.dart
@@ -64,6 +64,7 @@ void main() {
           linkModePreference: linkMode == DynamicLoadingBundled()
               ? LinkModePreference.dynamic
               : LinkModePreference.static,
+          linkingEnabled: false,
         );
         final buildOutput = BuildOutput();